move the OS-X/windows pkg-build/test commands back into master.cfg
I think bbsupport.py is way too big, the buildbot config as a whole is
too complicated, and I want to see more work being done by helper
scripts (shell and python) in the source trees. The new OS-X/windows
steps do just that, which is awesome.

So I think the buildmaster ShellCommands that drive them can live in
master.cfg instead of living with the more-complicated things in
bbsupport.py . Also, they reference helper scripts which live in
the *tahoe* source tree, and the new BuildSteps are only used once, both
of which are arguments in favor of inlining.

This also changes tahoe's make_factory() to define python= (as in
"provide the default in make_factory()", instead of "sometimes pass None
down into various builtsteps and let them provide the default"), which
simplifies things somewhat.

Note that the "test-windows-package.py" script doesn't exist yet. In
this patch, I added a call to another not-yet-existing script named
"build-windows-package.py".

The windows steps are currently being used by daira's buildslave, but no
buildslaves use the new Mac steps yet. I'll add one in a subsequent
patch.