remove memcheck-32, it got upgraded to a 64-bit kernel
The buildslave (warner-linode, aka foolscap.lothar.com) got upgraded
around 04-Feb-2016. I don't think we have a 32-bit machine around
anymore. The 64-bit machines show an absurdly high memory usage. We
measure /proc/self/status/VmSize, which is somewhat absurd anyways, but
it used to be a useful proxy for peak memory usage, and would indicate
when we were trying to hold the whole file in memory (instead of
streaming it properly).

This whole thing probably needs to be redesigned.